Bug corregido

+++ strings/kmp/kmp.cpp
@@ -1,7 +1,7 @@
 const int ALPHABET = 26;
 vector<int> CalculatePi(const string& s) {
   int n = s.size();
-  vector<Long> pi(n, 0);
+  vector<int> pi(n, 0);
   for (int i = 1; i < n; i++) {
     int j = pi[i - 1];
     while (j > 0 && s[i] != s[j]) j = pi[j - 1];
